init as contao 4 module
