+++ .gitignore
@@ -1,1 +1,18 @@
+# OS
 .DS_Store
+Thumbs.db
+
+# IDEs 
+.buildpath
+.project
+.settings/
+.build/
+.external*/
+.idea/
+nbproject/
+
+# composer related
+vendor/
+
+# build
+build/
