+++ src/Resources/contao/classes/ContentICal.php
