+++ src/Resources/contao/config/autoload.ini
