+++ src/Resources/contao/config/autoload.php
