+++ src/Resources/contao/languages/en/tl_calendar.php
