+++ src/Resources/contao/templates/ce_ical.html5
