+++ src/Resources/public/.htaccess
